feat: Renamed DatabaseInitializer.kt to DatabaseInit.kt and its method. Implemented enums for the cardtypes, landmarktypes etc. in domain folder
