+++ src/main/kotlin/org/machikoro/server/domain/CardType.kt
@@ -1,0 +1,19 @@
+package org.machikoro.server.domain
+
+enum class CardType {
+    WHEAT_FIELD,
+    RANCH,
+    FOREST,
+    MINE,
+    APPLE_ORCHARD,
+    BAKERY,
+    CONVENIENCE_STORE,
+    CHEESE_FACTORY,
+    FURNITURE_FACTORY,
+    FRUIT_AND_VEGETABLE_MARKET,
+    CAFE,
+    FAMILY_RESTAURANT,
+    STADIUM,
+    TV_STATION,
+    BUSINESS_CENTER,
+}
