+++ src/main/kotlin/org/machikoro/server/domain/GameStatus.kt
@@ -1,0 +1,7 @@
+package org.machikoro.server.domain
+
+enum class GameStatus {
+    WAITING,
+    IN_PROGRESS,
+    FINISHED
+}
