+++ src/main/kotlin/org/machikoro/server/domain/LandmarkType.kt
@@ -1,0 +1,8 @@
+package org.machikoro.server.domain
+
+enum class LandmarkType {
+    TRAIN_STATION,
+    SHOPPING_MALL,
+    AMUSEMENT_PARK,
+    RADIO_TOWER
+}
