+++ src/main/kotlin/org/machikoro/server/domain/TurnPhase.kt
@@ -1,0 +1,8 @@
+package org.machikoro.server.domain
+
+enum class TurnPhase {
+    ROLL_DICE,
+    RESOLVE_EFFECTS,
+    BUY_OR_BUILD,
+    END_TURN
+}
